+++ support/src/main/java/io/jenkins/plugins/casc/support/groovy/GroovyScriptSource.java
@@ -1,0 +1,14 @@
+package io.jenkins.plugins.casc.support.groovy;
+
+import hudson.ExtensionPoint;
+import hudson.model.AbstractDescribableImpl;
+
+import java.io.IOException;
+
+/**
+ * @author <a href="mailto:tomasz.szandala@gmail.com">Tomasz Szandala</a>
+ */
+public abstract class GroovyScriptSource extends AbstractDescribableImpl<GroovyScriptSource> implements ExtensionPoint {
+
+    public abstract String getScript() throws IOException;
+}