+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.4 (not released yet)
 - Add support for Vault appRole authentication method
+- Allow to run Groovy scripts as a part of configuration Yaml (inline or other source)
 
 ## 1.3
 
@@ -24,7 +25,7 @@
 ## 1.1
 
 - [SECURITY-1124] Never export sensitive Secret
-- fix plugin installation 
+- fix plugin installation
 - impersonate as SYSTEM to apply configuration
 - removed Beta API annotations
 - many fixes to documentation and demos